feat: authorization implemented

+++ module/notes-app-back-end/src/api/notes/handler.js
@@ -1,5 +1,5 @@
 // @ Request Handling
-const ClientError = require('../../exceptions/ClientError');
+const ClientError = require("../../exceptions/ClientError");
 
 //? digunakan untuk menangani permintaan dari client yang datang kemudian memberikan respons dan sebaiknya memang hanya sebatas itu
 class NotesHandler {
@@ -8,7 +8,7 @@ class NotesHandler {
     this._service = service;
     this._validator = validator;
 
-    //? binding each handler, for fix javascript behaviour 
+    //? binding each handler, for fix javascript behaviour
     this.postNoteHandler = this.postNoteHandler.bind(this);
     this.getNotesHandler = this.getNotesHandler.bind(this);
     this.getNoteByIdHandler = this.getNoteByIdHandler.bind(this);
@@ -20,8 +20,14 @@ class NotesHandler {
     try {
       this._validator.validateNotePayload(request.payload);
       const { title = "untitled", body, tags } = request.payload;
+      const { id: credentialId } = request.auth.credentials; //? get id auth
 
-      const noteId = await this._service.addNote({ title, body, tags }); //? call funtion for prosess addNote
+      const noteId = await this._service.addNote({
+        title,
+        body,
+        tags,
+        owner: credentialId,
+      }); //? call funtion for prosess addNote
 
       const response = h.response({
         status: "success",
@@ -35,17 +41,17 @@ class NotesHandler {
     } catch (error) {
       if (error instanceof ClientError) {
         const response = h.response({
-          status: 'fail',
+          status: "fail",
           message: error.message,
         });
         response.code(error.statusCode);
         return response;
       }
- 
+
       // Server ERROR!
       const response = h.response({
-        status: 'error',
-        message: 'Maaf, terjadi kegagalan pada server kami.',
+        status: "error",
+        message: "Maaf, terjadi kegagalan pada server kami.",
       });
       response.code(500);
       console.error(error);
@@ -53,8 +59,9 @@ class NotesHandler {
     }
   }
 
-  async getNotesHandler() {
-    const notes = await this._service.getNotes();
+  async getNotesHandler(request) {
+    const { id: credentialId } = request.auth.credentials;
+    const notes = await this._service.getNotes(credentialId);
     return {
       status: "success",
       data: {
@@ -63,9 +70,12 @@ class NotesHandler {
     };
   }
 
-  async getNoteByIdHandler(request) {
+  async getNoteByIdHandler(request, h) {
     try {
       const { id } = request.params; //? param from client
+      const { id: credentialId } = request.auth.credentials;
+
+      await this._service.verifyNoteOwner(id, credentialId); //? get notes from owner
       const note = await this._service.getNoteById(id);
       return {
         status: "success",
@@ -76,17 +86,17 @@ class NotesHandler {
     } catch (error) {
       if (error instanceof ClientError) {
         const response = h.response({
-          status: 'fail',
+          status: "fail",
           message: error.message,
         });
         response.code(error.statusCode);
         return response;
       }
- 
+
       // Server ERROR!
       const response = h.response({
-        status: 'error',
-        message: 'Maaf, terjadi kegagalan pada server kami.',
+        status: "error",
+        message: "Maaf, terjadi kegagalan pada server kami.",
       });
       response.code(500);
       console.error(error);
@@ -97,28 +107,30 @@ class NotesHandler {
   async putNoteByIdHandler(request, h) {
     try {
       this._validator.validateNotePayload(request.payload);
-      const { id } = request.params;  
- 
-      await this._service.editNoteById(id, request.payload);
- 
+      const { id } = request.params; //? id_note
+      const { id: credentialId } = request.auth.credentials; //? id_owner
+
+      await this._service.verifyNoteOwner(id, credentialId); //? cek note by id_owner
+      await this._service.editNoteById(id, request.payload); //? cek data by id_note
+
       return {
-        status: 'success',
-        message: 'Catatan berhasil diperbarui',
+        status: "success",
+        message: "Catatan berhasil diperbarui",
       };
     } catch (error) {
       if (error instanceof ClientError) {
         const response = h.response({
-          status: 'fail',
+          status: "fail",
           message: error.message,
         });
         response.code(error.statusCode);
         return response;
       }
- 
+
       // Server ERROR!
       const response = h.response({
-        status: 'error',
-        message: 'Maaf, terjadi kegagalan pada server kami.',
+        status: "error",
+        message: "Maaf, terjadi kegagalan pada server kami.",
       });
       response.code(500);
       console.error(error);
@@ -129,25 +141,29 @@ class NotesHandler {
   async deleteNoteByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+
+      await this._service.verifyNoteOwner(id, credentialId);
       await this._service.deleteNoteById(id);
+
       return {
-        status: 'success',
-        message: 'Catatan berhasil dihapus',
+        status: "success",
+        message: "Catatan berhasil dihapus",
       };
     } catch (error) {
       if (error instanceof ClientError) {
         const response = h.response({
-          status: 'fail',
+          status: "fail",
           message: error.message,
         });
         response.code(error.statusCode);
         return response;
       }
- 
+
       // Server ERROR!
       const response = h.response({
-        status: 'error',
-        message: 'Maaf, terjadi kegagalan pada server kami.',
+        status: "error",
+        message: "Maaf, terjadi kegagalan pada server kami.",
       });
       response.code(500);
       console.error(error);
